requirements.txt
beautifulsoup4
fastapi==0.86.0
requests==2.28.1
uvicorn==0.19.0
python-dotenv==0.21.0
